+++ applications/overlayConfig.js
@@ -106,7 +106,8 @@ export default class OverlayConfig extends FormApplication {
       showArtSelect(this.token?.name ?? 'overlay', {
         searchType: SEARCH_TYPE.TOKEN,
         callback: (imgSrc, imgName) => {
-          if (imgSrc) $(event.target).closest('.form-group').find('input').val(imgSrc).trigger('change');
+          if (imgSrc)
+            $(event.target).closest('.form-group').find('input').val(imgSrc).trigger('change');
         },
       });
     });
@@ -116,7 +117,11 @@ export default class OverlayConfig extends FormApplication {
       if (presetName) {
         const preset = TokenMagic.getPreset(presetName);
         if (preset) {
-          $(event.target).closest('.form-group').find('textarea').val(JSON.stringify(preset, null, 2)).trigger('input');
+          $(event.target)
+            .closest('.form-group')
+            .find('textarea')
+            .val(JSON.stringify(preset, null, 2))
+            .trigger('input');
         }
       }
     });
@@ -124,9 +129,13 @@ export default class OverlayConfig extends FormApplication {
     // Controls for locking scale sliders together
     let scaleState = { locked: true };
 
-    html.find('.offsetX, .offsetY').on('change', (event) => {
-      $(event.target).siblings('.range-value').val($(event.target).val()).trigger('change');
-    });
+    // Range inputs need to be triggered when slider moves to initiate preview
+    html
+      .find('.range-value')
+      .siblings('[type="range"]')
+      .on('change', (event) => {
+        $(event.target).siblings('.range-value').val(event.target.value).trigger('change');
+      });
 
     const lockButtons = $(html).find('.scaleLock > a');
     const sliderScaleWidth = $(html).find('[name="scaleX"]');
@@ -136,7 +145,9 @@ export default class OverlayConfig extends FormApplication {
 
     lockButtons.on('click', function () {
       scaleState.locked = !scaleState.locked;
-      lockButtons.html(scaleState.locked ? '<i class="fas fa-link"></i>' : '<i class="fas fa-unlink"></i>');
+      lockButtons.html(
+        scaleState.locked ? '<i class="fas fa-link"></i>' : '<i class="fas fa-unlink"></i>'
+      );
     });
 
     sliderScaleWidth.on('change', function () {
@@ -181,12 +192,14 @@ export default class OverlayConfig extends FormApplication {
         }
 
         if (param)
-          game.modules.get('multi-token-edit').api.showGenericForm(param, param.filterType ?? 'TMFX', {
-            inputChangeCallback: (selected) => {
-              mergeObject(param, selected, { inplace: true });
-              textarea.val(JSON.stringify(params, null, 2)).trigger('input');
-            },
-          });
+          game.modules
+            .get('multi-token-edit')
+            .api.showGenericForm(param, param.filterType ?? 'TMFX', {
+              inputChangeCallback: (selected) => {
+                mergeObject(param, selected, { inplace: true });
+                textarea.val(JSON.stringify(params, null, 2)).trigger('input');
+              },
+            });
       }
     });
 
@@ -428,7 +441,9 @@ export default class OverlayConfig extends FormApplication {
 
     data.fonts = Object.keys(CONFIG.fontDefinitions);
 
-    const allMappings = getAllEffectMappings(this.token, true).filter((m) => m.id !== this.config.id);
+    const allMappings = getAllEffectMappings(this.token, true).filter(
+      (m) => m.id !== this.config.id
+    );
     const [_, groupedMappings] = sortMappingsToGroups(allMappings);
 
     data.parents = groupedMappings;
@@ -461,6 +476,33 @@ export default class OverlayConfig extends FormApplication {
     return mergeObject(data, settings);
   }
 
+  _getHeaderButtons() {
+    const buttons = super._getHeaderButtons();
+    buttons.unshift({
+      label: 'Core Variables',
+      class: '.core-variables',
+      icon: 'fas fa-file-import fa-fw',
+      onclick: () => {
+        let content = `
+        <table>
+          <tr><th>Variable</th><th>Description</th></tr>
+          <tr><td>@hp</td><td>Actor Health</td></tr>
+          <tr><td>@hpMax</td><td>Actor Health (Max)</td></tr>
+          <tr><td>@gridSize</td><td>Grid Size (Pixels)</td></tr>
+          <tr><td>@label</td><td>Mapping's Label Field</td></tr>
+        </table>
+        `;
+
+        new Dialog({
+          title: `Core Variables`,
+          content,
+          buttons: {},
+        }).render(true);
+      },
+    });
+    return buttons;
+  }
+
   async close(options = {}) {
     super.close(options);
     this._removePreviews();
@@ -484,7 +526,8 @@ export default class OverlayConfig extends FormApplication {
       formData.variables = formData.variables.filter((v) => v.name.trim() && v.value.trim());
     }
     if (formData.limitedUsers) {
-      if (getType(formData.limitedUsers) === 'string') formData.limitedUsers = [formData.limitedUsers];
+      if (getType(formData.limitedUsers) === 'string')
+        formData.limitedUsers = [formData.limitedUsers];
       formData.limitedUsers = formData.limitedUsers.filter((uid) => uid);
     } else {
       formData.limitedUsers = [];
@@ -495,13 +538,21 @@ export default class OverlayConfig extends FormApplication {
     if (formData.parentID === 'TOKEN') formData.parentID = '';
 
     if (formData.filter === 'OutlineOverlayFilter' && 'filterOptions.outlineColor' in formData) {
-      formData['filterOptions.outlineColor'] = this._convertColor(formData['filterOptions.outlineColor']);
+      formData['filterOptions.outlineColor'] = this._convertColor(
+        formData['filterOptions.outlineColor']
+      );
     } else if (formData.filter === 'BevelFilter') {
       if ('filterOptions.lightColor' in formData)
-        formData['filterOptions.lightColor'] = Number(Color.fromString(formData['filterOptions.lightColor']));
+        formData['filterOptions.lightColor'] = Number(
+          Color.fromString(formData['filterOptions.lightColor'])
+        );
       if ('filterOptions.shadowColor' in formData)
-        formData['filterOptions.shadowColor'] = Number(Color.fromString(formData['filterOptions.shadowColor']));
-    } else if (['DropShadowFilter', 'GlowFilter', 'OutlineFilter', 'FilterFire'].includes(formData.filter)) {
+        formData['filterOptions.shadowColor'] = Number(
+          Color.fromString(formData['filterOptions.shadowColor'])
+        );
+    } else if (
+      ['DropShadowFilter', 'GlowFilter', 'OutlineFilter', 'FilterFire'].includes(formData.filter)
+    ) {
       if ('filterOptions.color' in formData)
         formData['filterOptions.color'] = Number(Color.fromString(formData['filterOptions.color']));
     }
@@ -1011,7 +1062,9 @@ function genControl(control, values) {
 <div class="form-group">
   <label>${label}</label>
   <div class="form-fields">
-      <input type="checkbox" name="filterOptions.${name}" data-dtype="Boolean" value="${val}" ${val ? 'checked' : ''}>
+      <input type="checkbox" name="filterOptions.${name}" data-dtype="Boolean" value="${val}" ${
+      val ? 'checked' : ''
+    }>
   </div>
 </div>
     `;
@@ -1024,7 +1077,9 @@ function genControl(control, values) {
 `;
 
     for (const opt of control.options) {
-      select += `<option value="${opt.value}" ${val === opt.value ? 'selected="selected"' : ''}>${opt.label}</option>`;
+      select += `<option value="${opt.value}" ${val === opt.value ? 'selected="selected"' : ''}>${
+        opt.label
+      }</option>`;
     }
 
     select += `</select></div></div>`;